backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.9.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Environment and configuration
python-dotenv==1.0.0

# Code execution and AI services
google-generativeai==0.3.2
google-genai==1.73.1

# Monitoring
prometheus-client>=0.20.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and logging
structlog==23.2.0

# Date and time utilities
python-dateutil==2.8.2

# JSON handling
orjson>=3.10.10

# Email services
fastapi-mail==1.4.1

# File upload and processing
aiofiles==23.2.1
pandas>=2.2.3
openpyxl==3.1.2

# Rate limiting
slowapi==0.1.9

# Health checks and Doc processing
psutil==5.9.6
PyPDF2==3.0.1
PyMuPDF==1.24.1
python-pptx==0.6.23
